+++ Reach/src/utopia/reach/component/input/selection/DropDown.scala
@@ -1,5 +1,6 @@
 package utopia.reach.component.input.selection
 
+import utopia.flow.operator.EqualsFunction
 import utopia.flow.view.immutable.View
 import utopia.flow.view.immutable.eventful.Fixed
 import utopia.flow.view.mutable.eventful.PointerWithEvents
@@ -91,7 +92,7 @@ case class ContextualDropDownFactory[+N <: TextContextLike](parentHierarchy: Com
 	 listLayout: StackLayout = Fit, listCap: StackLength = StackLength.fixedZero,
 	 noOptionsView: Option[OpenComponent[ReachComponentLike, Any]] = None,
 	 highlightStylePointer: Changing[Option[ColorRole]] = Fixed(None),
-	 focusColorRole: ColorRole = Secondary, sameItemCheck: Option[(A, A) => Boolean] = None,
+	 focusColorRole: ColorRole = Secondary, sameItemCheck: Option[EqualsFunction[A]] = None,
 	 fillBackground: Boolean = ComponentCreationDefaults.useFillStyleFields)
 	(makeDisplay: (ComponentHierarchy, A) => C)(implicit scrollingContext: ScrollingContextLike) =
 	{
@@ -161,7 +162,7 @@ case class ContextualDropDownFactory[+N <: TextContextLike](parentHierarchy: Com
 	 errorMessagePointer: Changing[LocalizedString] = Fixed(LocalizedString.empty),
 	 noOptionsView: Option[OpenComponent[ReachComponentLike, Any]] = None,
 	 highlightStylePointer: Changing[Option[ColorRole]] = Fixed(None), focusColorRole: ColorRole = Secondary,
-	 sameItemCheck: Option[(A, A) => Boolean] = None,
+	 sameItemCheck: Option[EqualsFunction[A]] = None,
 	 fillBackground: Boolean = ComponentCreationDefaults.useFillStyleFields)
 	(implicit scrollingContext: ScrollingContextLike) =
 	{